+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add AsciiDoc icon [Thanks safarli](https://github.com/BeardedBear/bearded-icons/pull/87)
 - Add new eslint config icon [Thanks lucasrmendonca](https://github.com/BeardedBear/bearded-icons/pull/88)
 - Add Motoko icon [Thanks LucasGuerraS](https://github.com/BeardedBear/bearded-icons/pull/91)
+- Add diff/patch files icon
+- Add quasar config icon
 
 ## 1.1.0 - 2022-09-21